* origin/develop: (47 commits)
  Grammar tweak
  get_stacktrace/0
  function_exported/3
  Increase min heap size to accomadate larger closures
  Don't use ? once code is non-reentrant
  Fix liblumen_eir_interpreter
  binary_to_term reimplemented
  Failing regression test for binary_to_term roundtrip
  Use explicit Ok(()) instead of return from unit function
  Use * for Copy types instead of clone
  Use assert_eq instead of assert(... == ...)
  Remove unnecessary clones
  Remove unnecessary return
  Remove useless as_ref
  Convert single match to if let
  Remove unnecessary let
  Lumen.Web.Window.window/0 native_implemented_function
  Lumen.Web.Window.document/1 native_implemented_function
  Lumen.Web.Window.add_event_listener/4 native_implemented_function
  Lumen.Web.Node.replace_child/3 native_implemented_function
  ...

+++ liblumen_alloc/src/erts/term/typed_term.rs
@@ -119,9 +119,9 @@ impl TypedTerm {
         }
     }
 
-    pub fn is_function_with_arity(&self, arity: usize) -> bool {
+    pub fn is_function_with_arity(&self, arity: u8) -> bool {
         match self {
-            Self::Closure(closure) => (closure.as_ref().arity() as usize) == arity,
+            Self::Closure(closure) => closure.as_ref().arity() == arity,
             _ => false,
         }
     }
